+++ src/types.ts
@@ -15,9 +15,10 @@ export enum Actions {
   APPROVE = 'approve',
   VERIFY = 'verify',
   REQUEST = 'request',
-  PURCHASE = 'purchase',
-  SHIP = 'ship',
-  RETURN = 'return'
+  // these values are uppercased to be compatible with keys to make it easier to access values without having to change the case --2019-08-06 @preeV42500
+  PURCHASE = 'PURCHASE',
+  SHIP = 'SHIP',
+  RETURN = 'RETURN'
 }
 export enum Resource {
   Division = 'Division',