+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
   * `moment` is no longer a peer dependency, just a `devDependency`
 ### Changed
+  * `Actions.PURCHASE`, `Actions.SHIP` and `Actions.RETURN` values are now uppercase
 ### Fixed
 ### Deprecated
 ### Security